Move tests to suite base. Create shared util function for job count.
