+++ src/test/scala/org/apache/spark/sql/delta/MergeIntoScalaSuite.scala
@@ -280,88 +280,6 @@ class MergeIntoScalaSuite extends MergeIntoSuiteBase {
     }
   }
 
-  test("findTouchedFiles - no files match target predicate") {
-    val jobCount = new AtomicInteger(0)
-    val listener = new SparkListener {
-      override def onJobStart(jobStart: SparkListenerJobStart): Unit = {
-        // Spark will always log a job start/end event even when the job does not launch any task.
-        if (jobStart.stageInfos.exists(_.numTasks > 0)) {
-          jobCount.incrementAndGet()
-        }
-      }
-    }
-    sparkContext.addSparkListener(listener)
-
-    withTable("source") {
-      append(Seq((1, 10), (2, 20)).toDF("key1", "value1"), "value1")  // target
-      val source = Seq((3, 30)).toDF("key2", "value2")  // source
-
-      try {
-        sparkContext.listenerBus.waitUntilEmpty(15000)
-        jobCount.set(0)
-
-        io.delta.tables.DeltaTable.forPath(spark, tempPath)
-          .merge(source, "key1 = key2 and value1 in (30)")
-          .whenMatched().updateExpr(Map("key1" -> "key2", "value1" -> "value2"))
-          .whenNotMatched().insertExpr(Map("key1" -> "key2", "value1" -> "value2"))
-          .execute()
-
-        sparkContext.listenerBus.waitUntilEmpty(15000)
-        assert(jobCount.get() == 3, "There should be 3 Spark jobs")
-
-        checkAnswer(
-          readDeltaTable(tempPath),
-          Row(1, 10) ::    // No change
-            Row(2, 20) ::     // No change
-            Row(3, 30) ::     // Insert
-            Nil)
-      } finally {
-        sparkContext.removeSparkListener(listener)
-      }
-    }
-  }
-
-  test("findTouchedFiles - files match target predicate") {
-    val jobCount = new AtomicInteger(0)
-    val listener = new SparkListener {
-      override def onJobStart(jobStart: SparkListenerJobStart): Unit = {
-        // Spark will always log a job start/end event even when the job does not launch any task.
-        if (jobStart.stageInfos.exists(_.numTasks > 0)) {
-          jobCount.incrementAndGet()
-        }
-      }
-    }
-    sparkContext.addSparkListener(listener)
-
-    withTable("source") {
-      append(Seq((1, 10), (2, 20)).toDF("key1", "value1"), "value1")  // target
-      val source = Seq((3, 30)).toDF("key2", "value2")  // source
-
-      try {
-        sparkContext.listenerBus.waitUntilEmpty(15000)
-        jobCount.set(0)
-
-        io.delta.tables.DeltaTable.forPath(spark, tempPath)
-          .merge(source, "key1 = key2")
-          .whenMatched().updateExpr(Map("key1" -> "key2", "value1" -> "value2"))
-          .whenNotMatched().insertExpr(Map("key1" -> "key2", "value1" -> "value2"))
-          .execute()
-
-        sparkContext.listenerBus.waitUntilEmpty(15000)
-        assert(jobCount.get() == 5, "There should be 5 Spark jobs")
-
-        checkAnswer(
-          readDeltaTable(tempPath),
-          Row(1, 10) ::    // No change
-            Row(2, 20) ::     // No change
-            Row(3, 30) ::     // Insert
-            Nil)
-      } finally {
-        sparkContext.removeSparkListener(listener)
-      }
-    }
-  }
-
   override protected def executeMerge(
       target: String,
       source: String,
